Mitigate time precision issues on ForceFieldPattern shader
On some mobile devices, a lower float precision used in the graphics
pipeline caused the ForceFieldPattern shader to distort in just a few
minutes, due to an high TIME value. This commit attempts to mitigate
this by using a modulo in the final alpha calculation, which causes
the shader to simply look laggy, but not distorted.