Give levels proper names, unrelated to their order
This will allow reordering levels more freely, without having to worry
about renaming them so that the names match the actual order.
It's still not perfect though, as it requires changing both the LEVEL_ORDER
const and the next_level variable in the relevant levels, so that the order
in the menu matches the order in which they're played.